refactor: Move KMeansResultsWidget to a new package
